requirements.txt
python-dotenv
requests
plexapi
gql[all]  # Installs gql with all supported transports
tqdm
